backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.18
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.34.0
torch==2.0.1
APScheduler==3.10.4
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
